refactor: Move `extractResource()` to `common-utils`
Prepare for wider reuse.
